requirements.txt
microservice-template-core==1.0.40
pymysql==1.0.2
flask_marshmallow==0.14.0
flask==1.1.2
marshmallow==3.10.0
werkzeug==0.16.1
itsdangerous==1.1.0
requests==2.25.1
urllib3==1.26.3
APScheduler==3.7.0
click==7.1.2
validators==0.18.2
watchdog==2.0.2
pyyaml==5.4.1